% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{box_cox}
\alias{box_cox}
\title{Box-Cox Transformation}
\usage{
box_cox(feature, lamda)
}
\arguments{
\item{feature}{is feature for transformation}

\item{lamda}{is conversion parameter}
}
\value{
transform feature
}
\description{
Box-Cox Transformation
}
