% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{classification_go}
\alias{classification_go}
\title{Classification}
\usage{
classification_go(data_study, data_test = NULL, model = "glmnet",
  y_name = "EndCat", is_optimal_border = FALSE, plot_roc = FALSE,
  feature_select = TRUE, multi = TRUE, anomalies = TRUE,
  box_cox = TRUE, feature_step = TRUE, stand = FALSE, n_folds = 0)
}
\arguments{
\item{data_study}{is data study}

\item{data_test}{is data test}

\item{model}{is name of model}

\item{y_name}{is a name of oucome colum}

\item{is_optimal_border}{is logical value using of method for optimal boarding}

\item{plot_roc}{is logical value is it necessary plot roc}

\item{feature_select}{is logical value is it necessary feature select}

\item{multi}{is logical value delete multicollinearity}

\item{anomalies}{is logical value using of method for optimal boarding}

\item{box_cox}{is logical value replace anomalies}

\item{feature_step}{is logical value select features by AUC}

\item{stand}{is logical value is it necessary feature standardization}

\item{n_folds}{is folds number}
}
\value{
list
}
\description{
Start classification
}
