% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clustering_go}
\alias{clustering_go}
\title{Get clusters}
\usage{
clustering_go(data, name_object, model = "Mclust", k = 3,
  prepare = FALSE, plot_cluster = TRUE, cluster_info = TRUE,
  box_plot = FALSE)
}
\arguments{
\item{data}{data for clustering}

\item{name_object}{is name of oject}

\item{model}{is model for clustering}

\item{k}{is count of clusters}

\item{prepare}{is logical value for prepare evidences before clustering}

\item{plot_cluster}{is logical value for plot cluster}

\item{cluster_info}{is logical value for get cluster information}

\item{box_plot}{is logical value for box plot cluster}
}
\value{
list
}
\description{
Start clustering
}
