% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{compare_classifications}
\alias{compare_classifications}
\title{Comparison classifications}
\usage{
compare_classifications(data_study, data_test = NULL, y_name = "y",
  models = c("gbm", "glm"), is_optimal_border = TRUE,
  feature_select = c(FALSE), multi = c(TRUE, FALSE),
  anomalies = c(TRUE, FALSE), box_cox = c(FALSE),
  feature_step = c(TRUE, FALSE), stand = TRUE, n_folds = 10,
  plot_test = FALSE)
}
\arguments{
\item{data_study}{data study}

\item{data_test}{data test}

\item{y_name}{is a name of oucome colum}

\item{models}{is name of models}

\item{is_optimal_border}{using of method for optimal boarding}

\item{feature_select}{is logical vector is it necessary feature select}

\item{multi}{is logical vector delete multicollinearity}

\item{anomalies}{is logical vector using of method for optimal boarding}

\item{box_cox}{is logical vector replace anomalies}

\item{feature_step}{is logical vector select features by AUC}

\item{stand}{logical vector is it necessary feature standardization}

\item{cv_n_folds}{folds number}
}
\value{
list
}
\description{
Comparison of several models
}
