% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{find_k}
\alias{find_k}
\title{find k}
\usage{
find_k(data, name_object, model = "kmeans", range_k = 2:30,
  prepare = TRUE)
}
\arguments{
\item{data}{data for clustering}

\item{name_object}{is name of oject}

\item{model}{is model for clustering}

\item{range_k}{is vector of counts of clusters}

\item{prepare}{is logical value for prepare evidences before clustering}
}
\value{
list
}
\description{
Get informationt about index for finding count clusters
}
