% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{get_box_plot}
\alias{get_box_plot}
\title{Plot box plot}
\usage{
get_box_plot(data, name_object, cluster)
}
\arguments{
\item{data}{data for clustering}

\item{name_object}{is name of oject}

\item{cluster}{is vector of cluster}
}
\value{
box plot
}
\description{
plot box plot clustering result
}
