% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{get_bss}
\alias{get_bss}
\title{Get bss}
\usage{
get_bss(x, cluster, center)
}
\arguments{
\item{x}{data for clustering}

\item{cluster}{is clustering result}

\item{center}{is center for cluster}
}
\value{
value
}
\description{
#' Get bss for clustering
}
