% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{get_cluster}
\alias{get_cluster}
\title{Get clusters}
\usage{
get_cluster(data, name_object, model = "kmeans", k = 5)
}
\arguments{
\item{data}{data for clustering}

\item{name_object}{is name of oject}

\item{model}{is model for clustering}

\item{k}{is count of clusters}
}
\value{
data frame
}
\description{
Get clusters
}
