% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{get_cluster_info}
\alias{get_cluster_info}
\title{Get cluster info}
\usage{
get_cluster_info(data, name_object, cluster)
}
\arguments{
\item{data}{data for clustering}

\item{name_object}{is name of oject}

\item{cluster}{is vector of cluster}
}
\value{
data frame
}
\description{
Get cluster info
}
