% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{get_data}
\alias{get_data}
\title{Get data}
\usage{
get_data(y_name, query, connection_string)
}
\arguments{
\item{y_name}{is name of outcome}

\item{query}{is query text}

\item{connectionStringDataMine}{is a connection string}
}
\value{
list of data and y_name
}
\description{
Get data of Sql Server
}
