% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{get_data_clust}
\alias{get_data_clust}
\title{Get data for clustering}
\usage{
get_data_clust(name_object = nULL, query, connection_string = nULL)
}
\arguments{
\item{name_object}{is name of oject}

\item{query}{is query text}

\item{connectionStringDataMine}{is a connection string}
}
\value{
list of data and y_name
}
\description{
Get data of Sql Server
}
