% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{get_feature_step}
\alias{get_feature_step}
\title{Features selection by AUC}
\usage{
get_feature_step(env, alg = "b")
}
\arguments{
\item{env}{is data environment}

\item{alg}{is algoritm of forward/backwards selection}
}
\value{
significant columns
}
\description{
Features selection by AUC
}
