% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{get_model}
\alias{get_model}
\title{Get model}
\usage{
get_model(data_study = NULL, model = "glm", y_name = "y")
}
\arguments{
\item{data_study}{data study}

\item{model}{is name of model}

\item{y_name}{is a name of oucome colum}
}
\value{
object
}
\description{
Comparison of several models
}
