% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{get_model_info}
\alias{get_model_info}
\title{model info}
\usage{
get_model_info(obj_model, model)
}
\arguments{
\item{obj_model}{model}

\item{model}{model name}
}
\value{
result data.frame
}
\description{
Get model informatoin
}
