% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{get_optimal_border}
\alias{get_optimal_border}
\title{Get optimal border}
\usage{
get_optimal_border(model_obj, data = NULL, model = "glm",
  y_name = "y")
}
\arguments{
\item{model_obj}{is model derived from classification_go}

\item{data}{data study}

\item{model}{is name of model}

\item{y_name}{is a name of oucome colum}
}
\value{
list
}
\description{
Get optimal border for get quality
}
