% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{get_plot}
\alias{get_plot}
\title{Get plot}
\usage{
get_plot(data, name_object, cluster, nameTest = "test", metod = "umap")
}
\arguments{
\item{data}{data for clustering}

\item{name_object}{is name of oject}

\item{cluster}{is vector of cluster}

\item{nameTest}{is name of plot}

\item{nameTest}{is metod for plot for count evidences > 3}
}
\value{
plot
}
\description{
Plot clustering result
}
