% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{get_prediction}
\alias{get_prediction}
\title{Get prediction}
\usage{
get_prediction(objModel, data = NULL, model = "glm", y_name = "y")
}
\arguments{
\item{objModel}{model object}

\item{data}{data for preidction}

\item{model}{is name of model}

\item{y_name}{is a name of oucome colum}
}
\value{
vecotr of prediction values
}
\description{
Prediction function for all methods
}
