% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R, R/clustering.R
\name{get_quality}
\alias{get_quality}
\title{Get quality classification}
\usage{
get_quality(data, name_object, cluster, center, k)

get_quality(data, name_object, cluster, center, k)
}
\arguments{
\item{data}{data for clustering}

\item{name_object}{is name of oject}

\item{cluster}{is vector of clusters}

\item{center}{is center for clusters}

\item{k}{is count of clusters}

\item{y}{true values for response}

\item{est}{estimation values for response}

\item{optimal_border}{using of method for optimal boarding}

\item{plot_roc}{is logical value for plot roc}
}
\value{
list of quality metrics

list
}
\description{
Get quality classification
}
