% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{get_quality_cv}
\alias{get_quality_cv}
\title{Cross-Validation}
\usage{
get_quality_cv(xy, n_folds = 10, model, y_name = "y",
  optimal_border = FALSE)
}
\arguments{
\item{xy}{data study}

\item{n_folds}{folds number}

\item{y_name}{is a name of oucome colum}

\item{optimal_border}{using of method for optimal boarding}

\item{models}{is name of model}
}
\value{
result data.frame
}
\description{
get quality cross-validation
}
