% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{get_wss}
\alias{get_wss}
\title{Get wss}
\usage{
get_wss(x, cluster, center)
}
\arguments{
\item{x}{data for clustering}

\item{cluster}{is clustering result}

\item{center}{is center for cluster}
}
\value{
value
}
\description{
#' Get wss for clustering
}
