% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{multi_features}
\alias{multi_features}
\title{Multicollinearity deletion}
\usage{
multi_features(env, method = "bartlett", alpha = 0.05, vif_bound = 4)
}
\arguments{
\item{env}{is data environment}

\item{method}{is choice of vif or bartlett}

\item{alpha}{is significant level}

\item{vif_bound}{is bound for vif}
}
\value{
remaining features
}
\description{
Multicollinearity deletion
}
