% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{prediction_go}
\alias{prediction_go}
\title{Prediction}
\usage{
prediction_go(data_predict, res_model)
}
\arguments{
\item{data_predict}{is data for prediction}

\item{res_model}{is model derived from classification_go}
}
\value{
result vector with predict
}
\description{
Start prediction
}
