% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{prepare_box_cox}
\alias{prepare_box_cox}
\title{Prepare data by box_cox}
\usage{
prepare_box_cox(env, lamda = seq(-4, 4, 0.01))
}
\arguments{
\item{env}{is data environment}

\item{lamda}{range for lamda}
}
\value{
data frame with information about lamda
}
\description{
Prepare data by box_cox transformation
}
