% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{prepare_evidences}
\alias{prepare_evidences}
\title{Prepare evidences before clustering}
\usage{
prepare_evidences(data, name_object)
}
\arguments{
\item{data}{data for clustering}

\item{name_object}{is name of oject}
}
\value{
plot
}
\description{
#' Prepare evidences before clustering
}
