% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{prepare_study}
\alias{prepare_study}
\title{Prepare data}
\usage{
prepare_study(data_study, data_test = NULL, y_name = "EndCat",
  feature_select = FALSE, multi = TRUE, anomalies = TRUE,
  box_cox = FALSE, stand = FALSE, feature_step = FALSE)
}
\arguments{
\item{data_study}{data study}

\item{data_test}{data test}

\item{y_name}{is a name of oucome colum}

\item{feature_select}{logical value is it necessary feature select}

\item{multi}{is logical value delete multicollinearity}

\item{anomalies}{is logical value using of method for optimal boarding}

\item{box_cox}{is logical value replace anomalies}

\item{stand}{logical value is it necessary feature standardization}

\item{feature_step}{is logical value select features by AUC}
}
\value{
list
}
\description{
Prepare data study and test
}
