% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{replace_anomalies}
\alias{replace_anomalies}
\title{Replace anomalies}
\usage{
replace_anomalies(env, k = 1.5, max_part_outlier = 0.2, step = 0.5)
}
\arguments{
\item{env}{is data environment}

\item{k}{is coefficient for box plot}

\item{max_part_outlier}{is max part outlier for replacement}

\item{step}{step for change k}
}
\value{
data frame with change information
}
\description{
Replace anomalies
}
