% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{significant_features}
\alias{significant_features}
\title{Features selection}
\usage{
significant_features(env, alpha = 0.1)
}
\arguments{
\item{env}{is data environment}

\item{alpha}{is significant level}
}
\value{
significant columns
}
\description{
Features selection for classification
}
