% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{testBartlett}
\alias{testBartlett}
\title{Bartletta test}
\usage{
testBartlett(x, alpha)
}
\arguments{
\item{x}{is matrix of features}

\item{alpha}{is significant level}
}
\value{
value of statisic
}
\description{
Bartletta test for multicollinearity
}
