% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{w_test}
\alias{w_test}
\title{Wilcoxon test}
\usage{
w_test(x1, x2)
}
\arguments{
\item{x1}{is first vecot}

\item{x2}{is second vector}
}
\value{
value of statisic
}
\description{
wilcoxon test for numeric vectors
}
