% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{z_test}
\alias{z_test}
\title{Z-test}
\usage{
z_test(x1, x2)
}
\arguments{
\item{x1}{is first vecot}

\item{x2}{is second vector}
}
\value{
value of statisic
}
\description{
#' Z-test for binary features
}
